pankegg as a one file python script
