--- bin/PANKEGG/__init__.py
@@ -1,4 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-from PANKEGG import *
