--- bin/PANKEGG/__main__.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-from PANKEGG.parser import main
-main()
